requirements.txt
playwright
chromadb
google-generativeai
python-dotenv
numpy
